% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytvdnrverify.R
\name{mytvdnrverify}
\alias{mytvdnrverify}
\title{Simple R-interface that calls the TVD-Nr verification}
\usage{
mytvdnrverify(psinputfile)
}
\arguments{
\item{psinputfile}{inputfile with TVD-numbers}
}
\description{
Inside of the inst/extdata we have a simple input file
containing TVD-Numbers to be verified.
}
