% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifyTvdNr.R
\docType{package}
\name{VerifyTvdNr}
\alias{VerifyTvdNr}
\alias{VerifyTvdNr-package}
\title{VerifyTvdNr Package To Test Verification of TVD-Numbers}
\description{
Given a TVD-Number, we generate an object of class SimpleAnimal
which specifies a TVD-Number as a string argument to the
constructor. The constructor calls a verification function
and returns the verified TVD-Number back.
}
