% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{verifyTvdNrMain}
\alias{verifyTvdNrMain}
\title{Testing Verification function of TVD-Numbers}
\usage{
verifyTvdNrMain(psinputFile)
}
\arguments{
\item{psinputFile}{name of an input file name}
}
\description{
An object of class simpleanimal is created where
the constructor calls a verification function.
This should verify that the specified TVD-Nr is
valid.
}
